+++ src/examples/index.ts
@@ -6,6 +6,7 @@ export * from './individual-controller';
 export * from './professional';
 export * from './employee';
 export * from './license';
+export * from './invoice';
 export * from './related-person';
 export * from './consent-access';
 export * from './relationship-access';